ai_auto_categorizer/requirements.txt
tensorflow
numpy
scikit-learn
Pillow
opencv-python
imagehash